+++ packages/components/alert-dialog/src/AlertDialog.stories.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react'
 import { AlertDialog } from '.'
 
 const meta: Meta<typeof AlertDialog> = {
-  title: 'Experimental/AlertDialog',
+  title: 'Components/AlertDialog',
   component: AlertDialog,
 }
 
@@ -49,58 +49,16 @@ export const Usage: StoryFn = _args => {
   )
 }
 
-export const Uncontrolled: StoryFn = () => {
-  return (
-    <AlertDialog defaultOpen={false}>
-      <AlertDialog.Trigger asChild>
-        <Button intent="danger">Delete</Button>
-      </AlertDialog.Trigger>
-
-      <AlertDialog.Portal>
-        <AlertDialog.Overlay />
-
-        <AlertDialog.Content>
-          <AlertDialog.Header>
-            <AlertDialog.Title>Delete account</AlertDialog.Title>
-          </AlertDialog.Header>
-
-          <AlertDialog.Body>
-            <AlertDialog.Description>
-              Are you sure? You can not undo this action afterwards.
-            </AlertDialog.Description>
-          </AlertDialog.Body>
-
-          <AlertDialog.Footer className="flex justify-end gap-md">
-            <AlertDialog.Cancel asChild>
-              <Button intent="neutral" design="ghost">
-                Cancel
-              </Button>
-            </AlertDialog.Cancel>
-
-            <AlertDialog.Action asChild>
-              <Button intent="danger">Delete</Button>
-            </AlertDialog.Action>
-          </AlertDialog.Footer>
-        </AlertDialog.Content>
-      </AlertDialog.Portal>
-    </AlertDialog>
-  )
-}
-
 export const Controlled: StoryFn = () => {
   const [isOpen, setIsOpen] = useState(false)
 
-  const handleClick = () => {
-    setIsOpen(true)
-  }
-
   return (
     <>
-      <Button intent="danger" onClick={handleClick}>
-        Delete
-      </Button>
-
       <AlertDialog open={isOpen} onOpenChange={setIsOpen}>
+        <AlertDialog.Trigger asChild>
+          <Button intent="danger">Delete</Button>
+        </AlertDialog.Trigger>
+
         <AlertDialog.Portal>
           <AlertDialog.Overlay />
 
@@ -110,9 +68,7 @@ export const Controlled: StoryFn = () => {
             </AlertDialog.Header>
 
             <AlertDialog.Body>
-              <AlertDialog.Description>
-                Are you sure? You can not undo this action afterwards.
-              </AlertDialog.Description>
+              Are you sure? You can not undo this action afterwards.
             </AlertDialog.Body>
 
             <AlertDialog.Footer className="flex justify-end gap-md">
@@ -130,134 +86,5 @@ export const Controlled: StoryFn = () => {
         </AlertDialog.Portal>
       </AlertDialog>
     </>
-  )
-}
-
-export const Escape: StoryFn = () => {
-  const handleEscapeKeyDown = (event: KeyboardEvent) => {
-    // optional
-    event.preventDefault()
-  }
-
-  return (
-    <AlertDialog>
-      <AlertDialog.Trigger asChild>
-        <Button intent="danger">Delete</Button>
-      </AlertDialog.Trigger>
-
-      <AlertDialog.Portal>
-        <AlertDialog.Overlay />
-
-        <AlertDialog.Content onEscapeKeyDown={handleEscapeKeyDown}>
-          <AlertDialog.Header>
-            <AlertDialog.Title>Delete account</AlertDialog.Title>
-          </AlertDialog.Header>
-
-          <AlertDialog.Body>
-            <AlertDialog.Description>
-              Are you sure? You can not undo this action afterwards.
-            </AlertDialog.Description>
-          </AlertDialog.Body>
-
-          <AlertDialog.Footer className="flex justify-end gap-md">
-            <AlertDialog.Cancel asChild>
-              <Button intent="neutral" design="ghost">
-                Cancel
-              </Button>
-            </AlertDialog.Cancel>
-
-            <AlertDialog.Action asChild>
-              <Button intent="danger">Delete</Button>
-            </AlertDialog.Action>
-          </AlertDialog.Footer>
-        </AlertDialog.Content>
-      </AlertDialog.Portal>
-    </AlertDialog>
-  )
-}
-
-export const OpenFocus: StoryFn = () => {
-  const handleOpenAutoFocus = (event: Event) => {
-    // optional
-    event.preventDefault()
-  }
-
-  return (
-    <AlertDialog>
-      <AlertDialog.Trigger asChild>
-        <Button intent="danger">Delete</Button>
-      </AlertDialog.Trigger>
-
-      <AlertDialog.Portal>
-        <AlertDialog.Overlay />
-
-        <AlertDialog.Content onOpenAutoFocus={handleOpenAutoFocus}>
-          <AlertDialog.Header>
-            <AlertDialog.Title>Delete account</AlertDialog.Title>
-          </AlertDialog.Header>
-
-          <AlertDialog.Body>
-            <AlertDialog.Description>
-              Are you sure? You can not undo this action afterwards.
-            </AlertDialog.Description>
-          </AlertDialog.Body>
-
-          <AlertDialog.Footer className="flex justify-end gap-md">
-            <AlertDialog.Cancel asChild>
-              <Button intent="neutral" design="ghost">
-                Cancel
-              </Button>
-            </AlertDialog.Cancel>
-
-            <AlertDialog.Action asChild>
-              <Button intent="danger">Delete</Button>
-            </AlertDialog.Action>
-          </AlertDialog.Footer>
-        </AlertDialog.Content>
-      </AlertDialog.Portal>
-    </AlertDialog>
-  )
-}
-
-export const CloseFocus: StoryFn = () => {
-  const handleCloseAutoFocus = (event: Event) => {
-    // optional
-    event.preventDefault()
-  }
-
-  return (
-    <AlertDialog>
-      <AlertDialog.Trigger asChild>
-        <Button intent="danger">Delete</Button>
-      </AlertDialog.Trigger>
-
-      <AlertDialog.Portal>
-        <AlertDialog.Overlay />
-
-        <AlertDialog.Content onCloseAutoFocus={handleCloseAutoFocus}>
-          <AlertDialog.Header>
-            <AlertDialog.Title>Delete account</AlertDialog.Title>
-          </AlertDialog.Header>
-
-          <AlertDialog.Body>
-            <AlertDialog.Description>
-              Are you sure? You can not undo this action afterwards.
-            </AlertDialog.Description>
-          </AlertDialog.Body>
-
-          <AlertDialog.Footer className="flex justify-end gap-md">
-            <AlertDialog.Cancel asChild>
-              <Button intent="neutral" design="ghost">
-                Cancel
-              </Button>
-            </AlertDialog.Cancel>
-
-            <AlertDialog.Action asChild>
-              <Button intent="danger">Delete</Button>
-            </AlertDialog.Action>
-          </AlertDialog.Footer>
-        </AlertDialog.Content>
-      </AlertDialog.Portal>
-    </AlertDialog>
   )
 }