+++ packages/components/dropdown/src/Dropdown.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react'
+
+import { DropdownProvider } from './DropdownContext'
+
+export interface DropdownProps {
+  children: ReactNode
+}
+
+export const Dropdown = ({ children }: DropdownProps) => {
+  return <DropdownProvider>{children}</DropdownProvider>
+}
+
+Dropdown.displayName = 'Dropdown'