+++ packages/components/dropdown/src/DropdownPopover.tsx
@@ -1,0 +1,30 @@
+import { Popover as SparkPopover } from '@spark-ui/popover'
+import { cx } from 'class-variance-authority'
+import { PropsWithChildren, useEffect } from 'react'
+
+import { useDropdown } from './DropdownContext'
+
+export const Popover = ({ children }: PropsWithChildren) => {
+  const { isOpen, hasPopover, setHasPopover } = useDropdown()
+
+  useEffect(() => {
+    setHasPopover(true)
+
+    return () => setHasPopover(false)
+  }, [setHasPopover])
+
+  if (!hasPopover) return children
+
+  return (
+    <SparkPopover.Content
+      matchTriggerWidth
+      onOpenAutoFocus={e => e.preventDefault()}
+      className={cx(!isOpen && 'hidden')}
+    >
+      {children}
+    </SparkPopover.Content>
+  )
+}
+
+Popover.id = 'Popover'
+Popover.displayName = 'Dropdown.Popover'