+++ packages/components/slider/src/SliderThumb.tsx
@@ -1,5 +1,5 @@
 import * as RadixSlider from '@radix-ui/react-slider'
-import { forwardRef } from 'react'
+import { type FocusEvent, forwardRef, type KeyboardEvent, type PointerEvent, useRef } from 'react'
 
 import { useSliderContext } from './SliderContext'
 import { thumbVariants } from './SliderThumb.styles'
@@ -12,14 +12,39 @@ export interface SliderThumbProps extends RadixSlider.SliderThumbProps {
   asChild?: boolean
 }
 
-export const SliderThumb = forwardRef<HTMLDivElement, SliderThumbProps>(
-  ({ asChild = false, className, ...rest }, ref) => {
+export const SliderThumb = forwardRef<HTMLSpanElement, SliderThumbProps>(
+  ({ asChild = false, className, onPointerDown, onKeyDown, onBlur, ...rest }, forwardedRef) => {
     const { intent } = useSliderContext()
+
+    const innerRef = useRef(null)
+    const ref = forwardedRef || innerRef
+
+    const setInteractionType = (e: KeyboardEvent | FocusEvent | PointerEvent) => {
+      /**
+       * Radix Slider implementation uses `.focus()` and thus prevent us to handle
+       * distinctively focus/focus-visible styles. So we use a `data-interaction` attribute to stay
+       * aware of the type of event, and adapt styles if needed.
+       */
+      if (typeof ref === 'function' || !ref.current) return
+      ref.current.dataset.interaction = e.type
+    }
 
     return (
       <RadixSlider.Thumb
         ref={ref}
         asChild={asChild}
+        onPointerDown={(e: PointerEvent<HTMLSpanElement>) => {
+          setInteractionType(e)
+          onPointerDown?.(e)
+        }}
+        onKeyDown={(e: KeyboardEvent<HTMLSpanElement>) => {
+          setInteractionType(e)
+          onKeyDown?.(e)
+        }}
+        onBlur={(e: FocusEvent<HTMLSpanElement>) => {
+          setInteractionType(e)
+          onBlur?.(e)
+        }}
         className={thumbVariants({ intent, className })}
         {...rest}
       />