+++ packages/components/checkbox/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.7](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@2.1.6...@spark-ui/checkbox@2.1.7) (2023-11-22)
+
+**Note:** Version bump only for package @spark-ui/checkbox
+
 ## [2.1.6](https://github.com/adevinta/spark/compare/@spark-ui/checkbox@2.1.5...@spark-ui/checkbox@2.1.6) (2023-10-30)
 
 **Note:** Version bump only for package @spark-ui/checkbox