+++ packages/components/switch/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.5](https://github.com/adevinta/spark/compare/@spark-ui/switch@2.2.4...@spark-ui/switch@2.2.5) (2023-11-23)
+
+**Note:** Version bump only for package @spark-ui/switch
+
+## [2.2.4](https://github.com/adevinta/spark/compare/@spark-ui/switch@2.2.3...@spark-ui/switch@2.2.4) (2023-11-22)
+
+**Note:** Version bump only for package @spark-ui/switch
+
 ## [2.2.3](https://github.com/adevinta/spark/compare/@spark-ui/switch@2.2.2...@spark-ui/switch@2.2.3) (2023-10-30)
 
 **Note:** Version bump only for package @spark-ui/switch